refactor: shared link service
